+++ src/main/java/com/lmsltirollcallsjtu/common/bean/bo/Students.java
@@ -7,7 +7,7 @@ import lombok.experimental.Tolerate;
 @Data
 @Builder
 public class Students {
-    private Long id;    //学生编号
+    private Integer id;    //学生编号
     private String name;//学生姓名
 
     @Tolerate
